Add ModuleSimplifier.
It simplifies the analysed modules by removing bodies of functions that are
same between modules. Function equivalence is determined by
llvm::FunctionComparator with custom numbering of global symbols to ensure
that corresponding globals get same number. This requires custom modification
of LLVM where GlobalNumberState is virtual.

The simplifier also creates abstractions of indirect function calls and of
inline assemblies.

Arguments of the printk function are removed.