Handle presence of two functions with different type of first parameter
This previously caused a wrong SMT formula to be produced in muz mode since
the parameter had the same name (_$1_0_old_1) but different type was expected.

Now, when removing forall expressions and collecting introduced variables, we
check whether a variable with same name and different type exists, and rename
the current one in such case.