Fix bug in call coupling in forbidden paths
If two calls to the same function occur in program, without a call to other
function in between, in two different basic blocks, there is wrong coupling of
the calls in forbidden paths. This bug occurs when adding forbidden paths where
one of the paths contains both calls whereas the other contains a single call.
In this case, since calls are coupled from the back, the only call in one path
is always coupled to the second of the two calls in the other path, which is
usually wrong.

This fixes uses the distance algorithm that is used to match the calls. If
calls corresponding to some row and column in the distance table are found to
be coupled, before matching them, we search for another coupled call in the
same row/column with lower distance and if such call is found, it has higher
precedence of matching.